requirements.txt
alembic==1.6.5
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.0.4
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.23
uritemplate==3.0.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
